Code/App_engine_model/Version_4_of_app_engine/requirements.txt
Flask
joblib
pandas
google-cloud-storage
gunicorn
google-cloud-logging
scikit-learn==1.1.3 